Collapsing ClickHouseBinaryReader into ClickHouseDataReader - no need for separate class
